unalias beer and better linux support

+++ bashrc.d/rails.bash
@@ -1,16 +1,9 @@
-alias sc='script/console'
-#alias sg='script/generate'
 #alias mig='rake db:migrate db:test:prepare'
-#alias tl='tail -f log/development.log'
+#alias tl='touch log/development.log ; tail -f log/development.log'
 #alias debug='bundle exec rdebug -c -no-stop'
-alias sat='bundle exec ~/bin/search_and_test.rb'
 alias gems='gem list'
 alias irb='irb --readline -r irb/completion'
-alias bake='bundle exec rake'
-alias beer='bundle exec ruby'
-alias cuke='RAILS_ENV=cucumber bundle exec cucumber'
-alias spec='ruby -Ispec'
-alias specr='spec --debug -Srdebug'
+
 #given a migration task, return the file name
 function mf() { ls db/migrate/*${1:?Please specify migration action}* ; }
 #given a migration task, return the version
@@ -18,7 +11,7 @@ function mf() { ls db/migrate/*${1:?Please specify migration action}* ; }
 function mver() { mf $1 | sed 's=^[^0-9]*\([0-9][^_]*\)_.*$=\1=' ; }
 
 function ruby-version {
-  local rv="ruby-1.9.3-p448"
+  local rv="ruby-2.2.3"
   if [ $# -eq 1 ] ; then
     chruby $1
     rv=`chruby | awk '/\*/ { print $2; }'`
@@ -29,6 +22,7 @@ function ruby-version {
 }
 
 function powrc {
+[ -f .powrc] && return
 cat <<EOF > .powrc
 source /usr/local/share/chruby/chruby.sh
 [ -f .ruby-version ] && chruby \$(cat .ruby-version)
@@ -38,7 +32,6 @@ EOF
 
 # Cucumber env setting 
 export AUTOFEATURE=true	
-#export RSPEC=true
 
 #ree settings
 #export RUBY_HEAP_MIN_SLOTS=600000
